+++ game/commander/tasks/primitive/antiship.py
@@ -13,6 +13,8 @@ from gen.flights.flight import FlightType
 @dataclass
 class PlanAntiShip(PackagePlanningTask[NavalGroundObject]):
     def preconditions_met(self, state: TheaterState) -> bool:
+        if not super().preconditions_met(state):
+            return False
         if self.target not in state.threatening_air_defenses:
             return False
         return self.target_area_preconditions_met(state, ignore_iads=True)