+++ game/commander/tasks/primitive/bai.py
@@ -12,6 +12,8 @@ from gen.flights.flight import FlightType
 @dataclass
 class PlanBai(PackagePlanningTask[VehicleGroupGroundObject]):
     def preconditions_met(self, state: TheaterState) -> bool:
+        if not super().preconditions_met(state):
+            return False
         if self.target not in state.enemy_garrisons:
             return False
         return self.target_area_preconditions_met(state)