+++ changelog.md
@@ -5,6 +5,7 @@ Saves from 3.x are not compatible with 5.0.
 ## Features/Improvements
 
 * **[Campaign AI]** Overhauled campaign AI target prioritization. This currently only affects the ordering of DEAD missions.
+* **[Campaign AI]** Player front line stances can now be automated. Improved stance selection for AI.
 
 ## Fixes